Move SVD classes into dedicated loader module
